Adds buffer checking to what4 proofs.
Basically, before this we were only checking the stream expressions, but now we
are also interpreting the first k timesteps concretely, checking that the
properties hold for all those specific indices. k is equal to the maximum buffer
length of all streams in the spec. External variables are the only unknowns, and
they are still represented as fresh constants.